Switch to Symfony Messenger component
The ServiceBus component and Bundle are completely gone now, the Symfony Messenger
provides everything we need (and more). In addition to these changes the Domain events
now use a dedicated bus rather then the Symfony EventListener.

Note: The busses are now shared by all modules, there are no separate command busses anymore.
The Command and Query bus are still separated from each other though.

The GuardMiddleware is replaced with a SecurityMiddleware that uses the Symfony security
system. For specific situations we use voters (with ServiceSubscribers to prevent eager loading).

Last, there is no adapter usage anymore for the busses.